auth/requirements.txt
fastapi-users[sqlalchemy]
aiosqlite
casbin
casbin_sqlalchemy_adapter
